move fits handling to separate module
